Remove `main.js`, inline into `index.js`
