+++ index.js
@@ -11,9 +11,13 @@
 import os from "os";
 import process from "process";
 
-import { escapeShellArg, quoteShellArg } from "./src/main.js";
+import { resolveExecutable } from "./src/executables.js";
 import { getHelpersByPlatform } from "./src/platforms.js";
-import { toArrayIfNecessary } from "./src/reflection.js";
+import {
+  checkedToString,
+  isString,
+  toArrayIfNecessary,
+} from "./src/reflection.js";
 
 /**
  * Get the helper functions for the current platform.
@@ -24,6 +28,33 @@ function getPlatformHelpers() {
   const platform = os.platform();
   const helpers = getHelpersByPlatform({ env: process.env, platform });
   return helpers;
+}
+
+/**
+ * Parses options provided to shescape.
+ *
+ * @param {object} args The arguments for this function.
+ * @param {object} args.options The options for escaping.
+ * @param {boolean} [args.options.flagProtection] Is flag protection enabled.
+ * @param {boolean} [args.options.interpolation] Is interpolation enabled.
+ * @param {boolean | string} [args.options.shell] The shell to escape for.
+ * @param {object} args.process The `process` values.
+ * @param {object} args.process.env The environment variables.
+ * @param {object} deps The dependencies for this function.
+ * @param {Function} deps.getDefaultShell Function to get the default shell.
+ * @param {Function} deps.getShellName Function to get the name of a shell.
+ * @returns {object} The parsed arguments.
+ */
+function parseOptions(
+  { options: { flagProtection, interpolation, shell }, process: { env } },
+  { getDefaultShell, getShellName }
+) {
+  flagProtection = flagProtection ? true : false;
+  interpolation = interpolation ? true : false;
+  shell = isString(shell) ? shell : getDefaultShell({ env });
+
+  const shellName = getShellName({ shell }, { resolveExecutable });
+  return { flagProtection, interpolation, shellName };
 }
 
 /**
@@ -52,7 +83,19 @@ function getPlatformHelpers() {
  */
 export function escape(arg, options = {}) {
   const helpers = getPlatformHelpers();
-  return escapeShellArg({ arg, options, process }, helpers);
+  const { shellName, interpolation, flagProtection } = parseOptions(
+    { options, process },
+    helpers
+  );
+  const argAsString = checkedToString(arg);
+  const escape = helpers.getEscapeFunction(shellName, { interpolation });
+  const escapedArg = escape(argAsString);
+  if (flagProtection) {
+    const flagProtect = helpers.getFlagProtectionFunction(shellName);
+    return flagProtect(escapedArg);
+  } else {
+    return escapedArg;
+  }
 }
 
 /**
@@ -116,7 +159,19 @@ export function escapeAll(args, options = {}) {
  */
 export function quote(arg, options = {}) {
   const helpers = getPlatformHelpers();
-  return quoteShellArg({ arg, options, process }, helpers);
+  const { flagProtection, shellName } = parseOptions(
+    { options, process },
+    helpers
+  );
+  const argAsString = checkedToString(arg);
+  const [escape, quote] = helpers.getQuoteFunction(shellName);
+  const escapedArg = escape(argAsString);
+  if (flagProtection) {
+    const flagProtect = helpers.getFlagProtectionFunction(shellName);
+    return quote(flagProtect(escapedArg));
+  } else {
+    return quote(escapedArg);
+  }
 }
 
 /**
