--- test/unit/main/_.js
@@ -1,10 +1,0 @@
-/**
- * @overview Provides testing utilities.
- * @license MIT
- */
-
-import * as arbitrary from "../../_arbitraries.js";
-import * as constants from "../../_constants.cjs";
-import * as macros from "./_macros.js";
-
-export { arbitrary, constants, macros };
